+++ src/classes/CodeGenerators/aggregation/IAggregationGenerator.ts
@@ -1,4 +1,4 @@
-import { SAPNode } from "../../SAPNode";
+import { SAPNode } from "../../StandardLibMetadata/SAPNode";
 
 export interface IAggregationGenerator {
 	generateAggregations(node: SAPNode) : Promise<string>;
