+++ src/classes/CodeGenerators/property/IPropertyGenerator.ts
@@ -1,4 +1,4 @@
-import { SAPNode } from "../../SAPNode";
+import { SAPNode } from "../../StandardLibMetadata/SAPNode";
 
 export interface IPropertyGenerator {
 	generateProperties(node: SAPNode) : Promise<string>;
