+++ src/classes/DAOAndFactories/UI5MetadataDAO.ts
@@ -1,5 +1,5 @@
-import { UI5Metadata } from "../UI5Metadata";
-import { SAPNode } from "../SAPNode";
+import { UI5Metadata } from "../StandardLibMetadata/UI5Metadata";
+import { SAPNode } from "../StandardLibMetadata/SAPNode";
 import * as rp from "request-promise";
 import * as vscode from "vscode";
 import * as fs from "fs";
